requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web App
streamlit>=1.28.0

# Logging and Monitoring
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
